Backend/Models/suggestionapiconctd/requirements_fullstack.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.32.5
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.36
alembic==1.14.0
httpx==0.28.1
pytest==8.3.4
